fixed comment download
