+++ js/form-validation.js
@@ -1,8 +1,5 @@
 import {isStringTooLong} from './util.js';
-import {renderSuccessMessage, renderErrorMessage} from './message.js';
-import {closeUploadModal} from './form-control.js';
 
-const uploadForm = document.querySelector('.img-upload__form');
 const hashtagInput = document.querySelector('.text__hashtags');
 const commentInput = document.querySelector('.text__description');
 
@@ -49,32 +46,4 @@ commentInput.addEventListener('input', () => {
   commentInput.reportValidity();
 });
 
-const formSubmit = (onSuccess) => {
-  uploadForm.addEventListener('submit', (evt) => {
-    evt.preventDefault;
-
-    const formData = new FormData(evt.target);
-
-    fetch('https://22.javascript.pages.academy/kekstagram',
-      {
-        method: 'POST',
-        body: formData,
-      },
-    )
-      .then((response) => {
-        if(response.ok) {
-          renderSuccessMessage();
-          onSuccess();
-        } else {
-          renderErrorMessage();
-          closeUploadModal();
-        }
-      })
-      .catch(() => {
-        renderErrorMessage();
-        closeUploadModal();
-      })
-  });
-};
-
-export {hashtagInput, commentInput, uploadForm, formSubmit};
+export {hashtagInput, commentInput };
